app/requirements.txt
streamlit
python-dotenv
langchain
PyPDF2
faiss-cpu
langchain-openai
langchain-community